requirements.txt
# QA Office — Phase 1 dependencies
# These are already installed as part of qa-os; listed here for standalone installs.

# Core
playwright>=1.40.0
openpyxl>=3.1.0
Pillow>=10.0.0
anthropic>=0.44.0  # 0.18.x fails on Python 3.12 due to httpx proxies kwarg removal
openai>=1.30.0
requests>=2.31.0
chromadb>=0.4.0

# Config
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
